Fix OpenSSL builds with MSVC.
OpenSSL is doind clang-cl detection by calling `cl.exe` with certain
arguments that aren't valid for MS cl.exe.  I'm not sure when this
behaviour started.  Just as happens with Boost, the instance of msvc
that runs the ExternalProject defines VS_UNICODE_OUTPUT to redirect all
output to a pipe.  That means is sees these non-fatal errors that are
happening all the way down in the OpenSSL configure step, and thinks the
build has failed.

This change works around the problem in the same way as the Boost scheme
by setting LOG_CONFIGURE, which wraps the ExternalProject build in a
script that insulates it from msvc's environment variables.