Fix logic in downloading of cache meta information
'x_DONE' file used as an indicator of successful operation. However if download
of 'x_DONE' finished but downloading of 'x_LOCAL' interrupted by user
the 'REMOVE x_DONE' is not called. This leads to invalid state when 'x_DONE'
present but 'x_LOCAL' not. Solution is to use temporary files. Files will
be renamed after all downloads and error checks.